+++ lib/screens/examples/multiselect_menu_button_example2.dart
@@ -7,9 +7,9 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 /// No issues
 ///
 
-final providerOfCitySelection =
-StateNotifierProvider<SelectionNotifier, SelectionModel>(
-        (ref) => SelectionNotifier(ref));
+final providerOfCitySelection = StateNotifierProvider<
+    MultipleSelectionNotifier<String>,
+    MultipleSelectionModel<String>>((ref) => MultipleSelectionNotifier(ref));
 
 final providerOfCities = FutureProvider((ref) async {
   await Future.delayed(const Duration(seconds: 1));
@@ -33,83 +33,92 @@ final providerOfCities = FutureProvider((ref) async {
   return choices;
 });
 
-enum SelectionState {
+enum MultipleSelectionState {
   all('(All)'),
   none('(None)'),
   some('(Some)');
 
-  const SelectionState(this._value);
+  const MultipleSelectionState(this._value);
   final String _value;
 
   @override
   String toString() => _value;
 
-  SelectionState parse(String value) {
+  MultipleSelectionState parse(String value) {
     return switch (value) {
-      '(All)' => SelectionState.all,
-      '(None)' => SelectionState.none,
-      '(Some)' => SelectionState.some,
+      '(All)' => MultipleSelectionState.all,
+      '(None)' => MultipleSelectionState.none,
+      '(Some)' => MultipleSelectionState.some,
       _ => throw ArgumentError('Invalid SelectionState $value'),
     };
   }
 }
 
-class SelectionModel {
-  SelectionModel({required this.selection, required this.choices});
+class MultipleSelectionModel<K> {
+  MultipleSelectionModel({required this.selection, required this.choices});
 
-  late final Set<String> selection;
-  final Set<String> choices;
+  late final Set<K> selection;
+  final Set<K> choices;
 
-  SelectionState get selectionState {
-    if (selection.isEmpty) return SelectionState.none;
+  MultipleSelectionState get selectionState {
+    if (selection.isEmpty) return MultipleSelectionState.none;
     if (choices.difference(selection).isNotEmpty) {
-      return SelectionState.some;
+      return MultipleSelectionState.some;
     } else {
-      return SelectionState.all;
+      return MultipleSelectionState.all;
     }
   }
 
-  SelectionModel add(String value) {
-    return SelectionModel(selection: selection..add(value), choices: choices);
+  MultipleSelectionModel<K> add(K value) {
+    return MultipleSelectionModel(
+        selection: selection..add(value), choices: choices);
   }
 
-  SelectionModel remove(String value) {
+  MultipleSelectionModel<K> remove(K value) {
     selection.remove(value);
-    return SelectionModel(selection: selection, choices: choices);
+    return MultipleSelectionModel(selection: selection, choices: choices);
   }
 
-  SelectionModel selectAll() {
-    return SelectionModel(selection: {...choices}, choices: choices);
+  MultipleSelectionModel<K> selectAll() {
+    return MultipleSelectionModel(selection: {...choices}, choices: choices);
   }
 
-  SelectionModel selectNone() {
-    return SelectionModel(selection: <String>{}, choices: choices);
+  MultipleSelectionModel<K> selectNone() {
+    return MultipleSelectionModel(selection: <K>{}, choices: choices);
   }
 
-  SelectionModel copyWith({Set<String>? selection, Set<String>? choices}) {
-    return SelectionModel(
+  MultipleSelectionModel<K> clone() => MultipleSelectionModel(
+      selection: <K>{...selection}, choices: <K>{...choices});
+
+  MultipleSelectionModel<K> copyWith({Set<K>? selection, Set<K>? choices}) {
+    return MultipleSelectionModel(
         selection: selection ?? this.selection,
         choices: choices ?? this.choices);
   }
 }
 
-class SelectionNotifier extends StateNotifier<SelectionModel> {
-  SelectionNotifier(this.ref)
-      : super(SelectionModel(selection: <String>{}, choices: <String>{}));
+class MultipleSelectionNotifier<K>
+    extends StateNotifier<MultipleSelectionModel<K>> {
+  MultipleSelectionNotifier(this.ref)
+      : super(MultipleSelectionModel(selection: <K>{}, choices: <K>{}));
   final Ref ref;
-  set add(String value) {
+  set add(K value) {
     state = state.add(value);
   }
+
   void selectAll() {
     state = state.selectAll();
   }
-  set remove(String value) {
+
+  set remove(K value) {
     state = state.remove(value);
   }
+
   void selectNone() {
     state = state.selectNone();
   }
-  void init(Set<String> selection, Set<String> choices) {
+
+  void init(Set<K> selection, Set<K> choices) {
     state = state.copyWith(selection: selection, choices: choices);
   }
 }
@@ -129,11 +138,11 @@ class _DropdownExampleState
   List<PopupMenuItem<String>> getList() {
     var model = ref.watch(providerOfCitySelection);
     var out = <PopupMenuItem<String>>[];
-    if (model.selectionState == SelectionState.all) {
+    if (model.selectionState == MultipleSelectionState.all) {
       setState(() {
         ref.read(providerOfCitySelection.notifier).selectAll();
       });
-    } else if (model.selectionState == SelectionState.none) {
+    } else if (model.selectionState == MultipleSelectionState.none) {
       setState(() {
         ref.read(providerOfCitySelection.notifier).selectNone();
       });
@@ -145,9 +154,12 @@ class _DropdownExampleState
         builder: (context, ref, child) {
           var model = ref.watch(providerOfCitySelection);
           return CheckboxListTile(
-            value: model.selectionState == SelectionState.all,
+            value: model.selectionState == MultipleSelectionState.all,
             controlAffinity: ListTileControlAffinity.leading,
-            title: const Text('(All)', style: TextStyle(fontFamily: 'Ubuntu'),),
+            title: const Text(
+              '(All)',
+              style: TextStyle(fontFamily: 'Ubuntu'),
+            ),
             onChanged: (bool? checked) {
               setState(() {
                 if (checked!) {
@@ -229,11 +241,16 @@ class _DropdownExampleState
                                 padding: const EdgeInsets.all(8.0),
                                 child: Row(
                                   children: [
-                                    Text(model.selectionState.toString(),
-                                      style: const TextStyle(fontFamily: 'Ubuntu', fontSize: 16),),
+                                    Text(
+                                      model.selectionState.toString(),
+                                      style: const TextStyle(
+                                          fontFamily: 'Ubuntu', fontSize: 16),
+                                    ),
                                     const Spacer(),
                                     const Icon(
-                                      Icons.keyboard_arrow_down_outlined, color: Colors.blueGrey,),
+                                      Icons.keyboard_arrow_down_outlined,
+                                      color: Colors.blueGrey,
+                                    ),
                                   ],
                                 ),
                               );